requirements.txt
djangorestframework
pyrankvote

Django
gunicorn
psycopg2-binary
PyYAML
docker-py
dockerpty
docopt
requests
six